Switch from InMemory to FromCode files.
I was concerned that I was not getting data appropriately raw from STDOUT from
an external process. It was *probably* correct, but I never liked setting
`binmode` on STDOUT, since that could muck things up for other parts of dzil.
So have `msgfmt` write to a file in a temporary directory, instead, collect
tha contents of that file myself as part of a FromCode file, and delete the
temporary directory in `DEMOLISH`. Fortunately, dzil writes the data back out
using `:raw`, so we should be golden now.